sample_segmentation_project/requirements.txt
# Core ML/DL libraries
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0

# Data processing and visualization
opencv-python>=4.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0

# Image processing and augmentation
albumentations>=1.1.0
scikit-image>=0.18.0

# Metrics and evaluation
scikit-learn>=1.0.0

# Utilities
PyYAML>=5.4.0
tensorboard>=2.7.0

# Optional: for better performance
# cupy-cuda11x  # If using CUDA 11.x
# cupy-cuda12x  # If using CUDA 12.x
